% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_metero.R
\docType{data}
\name{cal_metero}
\alias{cal_metero}
\title{Daily rainfall and evapotranspiration data for Caloosahatchee River}
\format{
An object of class \code{data.frame} with 17166 rows and 4 columns.
}
\source{
Qiu and Wan (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415
}
\usage{
data(cal_metero)
}
\description{
Rainfall and evapotranspiration data (in inches) presented in Qiu and Wan et al (2013) used to estimate tidal basin discharge to an ungauge system.
}
\references{
Qiu and Wan (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415
}
\keyword{dataset}
