% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_Q.R
\name{cost_Q}
\alias{cost_Q}
\title{Coastal Rainfall-Runoff discharge model (LinRes)}
\usage{
cost_Q(
  RF_2d,
  ETPI,
  RootStor = 1,
  Kc = 1,
  Z1.Smax = 1,
  Z1.StoreCef = 150,
  BasinArea = 251583,
  Z2.Smax = 1,
  Z2.StoreCef = 100,
  Z3.Smax = 2.2,
  Z3.StoreCef = 12,
  open.water.area = 0
)
}
\arguments{
\item{RF_2d}{2-day moving average rainfall value for the Caloosahatchee Tidal Basin (current forecast models use SFWMD NEXRAD values)}

\item{ETPI}{Regional evapotranspiration potential value for the (current forecast models use SFWMD DBKEY OH520)}

\item{RootStor}{Capacity of root zone storage (i.e. maximum root zone storage). Default is 1 inch.}

\item{Kc}{ET crop correction coefficient. Default is 1.}

\item{Z1.Smax}{Capacity of subsurface (Zone 1) drainage storage. Default is 1 inch.}

\item{Z1.StoreCef}{Storage coefficient of subsurface (Zone 1) drainage reservoir Default is 150.}

\item{BasinArea}{Area of tidal basin. Caloosahatchee tidal basin area is 251,583 acres.}

\item{Z2.Smax}{Capacity of surface (Zone 2) drainage storage. Default is 1 inch.}

\item{Z2.StoreCef}{Storage coefficient of surface (Zone 2) drainage reservoir Default is 100.}

\item{Z3.Smax}{Capacity of direct (Zone 3) drainage storage. Default is 2.2 inch.}

\item{Z3.StoreCef}{Storage coefficient of direct (Zone 3) drainage reservoir Default is 12.}

\item{open.water.area}{area of open water to estimate direct rainfall contribution.}
}
\value{
a vector of daily discharge values in cubic feet per second
}
\description{
Tidal Basin rainfall-runoff model as presented by Wan and Konyha (2015)
}
\examples{
data(cal_metero)
with(cal_metero,cost_Q(RF_2d,ETPI))
}
\references{
Wan Y, Konyha K (2015) A simple hydrologic model for rapid prediction of runoff from ungauged coastal catchments. Journal of Hydrology 528:571–583. doi: 10.1016/j.jhydrol.2015.06.047
}
