% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalMod.R
\name{SalMod}
\alias{SalMod}
\title{Salinity model as used by Qiu and Wan (2013)}
\usage{
SalMod(Q.total, SITE = "Val-I75", a.val = NULL, b.val = NULL, Q.thres = NULL)
}
\arguments{
\item{Q.total}{total discharge from freshwater inputs. For Caloosahatchee Estuary this includes tidal basin discharge and}

\item{SITE}{Current mointoring location with estalished salinity relationships. Sites include S79, BR31, Val-I75 and FtMyers. See Qiu and Qan (2013).}

\item{a.val}{alpha value model coefficient}

\item{b.val}{beta value model coefficient}

\item{Q.thres}{discharge threshold}
}
\value{
modeled salinity based on model coefficients and total freshwater discharge
}
\description{
Salinity model as used by Qiu and Wan (2013)
}
\examples{
sim.Q=c(500, 1001, 1500, 1998, 2507, 2994, 3498, 4006, 4511, 4991)

SalMod(Q.total=sim.Q)
}
\references{
Qiu C, Wan Y (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415
}
