% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteSalConstant.R
\docType{data}
\name{SiteSalConstant}
\alias{SiteSalConstant}
\title{Site specific salinity model constants for sites within the Caloosahatchee Estuary}
\format{
An object of class \code{data.frame} with 5 rows and 6 columns.
}
\source{
Qiu and Wan (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415
}
\usage{
data(cal_metero)
}
\description{
Site name, distance from S-79 and salinity model constants for sites within the Caloosahatchee Estuary. For more detail see Qiu and Wan (2013).
}
\references{
Qiu and Wan (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415
}
\keyword{dataset}
